Separated binary and library code
