--- src/lib.rs
@@ -1,6 +1,0 @@
-pub mod chip8 {
-    pub mod constants;
-    pub mod cpu;
-    mod handlers;
-    mod instructions;
-}
