+++ src/lib/lib.rs
@@ -1,0 +1,2 @@
+pub mod constants;
+pub mod cpu;
